BLE: Prevent writing before the subscription has completed
Apparantly the CoreBluetooth implementation and/or timing differs between macOS and iOS.
On macOS, it takes much longer to subscribe to value changes than on iOS, hence we sent
something down the output stream before making sure that we can actually pick up the responses.

It's now reworked in a way that the input channel only reports being `open` _after_ we
got a successful confirmation from the CoreBluetooth layer. Moreover, we now find out when
the subscription fails (for whatever reason) and signalize an error upwards the call chain.