FileHandleInputStream: fix reporting hasBytesAvailable and maximize reading block (where possible).
Previously, we always returned `true` in `hasBytesAvailable`. This is ok as long as the clients always
wait for the stream delegate to fire. Clients trying to immediately read after open would fail in that case.

On Linux, we now read up to the maximum allowed buffer size as indicated by the client. On Darwin, this
fails with an infamous _Encountered read failure 35 Resource temporarily unavailable_, hence we have to ifdef here.