TTYConnectionProvider: Provide a way to configure the baudrate of the attached device.
In contrast to other operating systems, this is a bit more involved on BSD and macOS,
as we need to keep the file descriptor around while using the device.

This is one of those situations where Linux has the non-standard behaviour,
but because it is so popular, people think there is some bug in the Mac.
In OS X and other versions of UNIX, closing the serial port restores it to its default settings
after the corresponding file descriptor closes.

Linux does not do that, hence `stty` is a bit more useful.